dev: move k8s primitives to k8s
