+++ pkg/model/k8s/service.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package primitives
+package k8s
 
 import (
 	"fmt"
