+++ pkg/model/k8s/stateful_set.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package primitives
+package k8s
 
 import (
 	apps "k8s.io/api/apps/v1"
